refactor(indexer): move streamer assembly into near-client
