+++ chain/indexer/Cargo.toml
@@ -13,7 +13,6 @@ workspace = true
 
 [dependencies]
 anyhow.workspace = true
-futures.workspace = true
 rocksdb.workspace = true
 tokio.workspace = true
 tracing.workspace = true
@@ -28,10 +27,8 @@ near-dyn-configs.workspace = true
 near-epoch-manager.workspace = true
 near-indexer-primitives.workspace = true
 near-o11y.workspace = true
-near-parameters.workspace = true
 near-primitives.workspace = true
 near-store.workspace = true
-node-runtime.workspace = true
 
 [features]
 calimero_zero_storage = ["near-primitives/calimero_zero_storage"]
@@ -42,11 +39,9 @@ nightly = [
     "near-epoch-manager/nightly",
     "near-indexer-primitives/nightly",
     "near-o11y/nightly",
-    "near-parameters/nightly",
     "near-primitives/nightly",
     "near-store/nightly",
     "nearcore/nightly",
-    "node-runtime/nightly",
 ]
 protocol_feature_spice = [
   "near-client/protocol_feature_spice",
